New version, changes mostly for imagery
This is really work in progress, I have used a lot already but some parts way
not be cleaned up. I have uploaded mostly so that users with scripting skills 
can use and modify it to their needs.
A second commit will follow just after to included examples of layers / extents / color_filters.