Fix some issues around earth window batch processing
There were a few troubles reading the per-tile config file from the
batch processing mode in the earth window.  This fixes a data type
on one of the attributes, a function name typo, and handling of
list types on load.

It appears to work in my testing so far, but the code could use a
review / sanity check from someone who understands it better than
I do, before being merged.