Auto-computes ZL zones around airports in Custom_Zoomlevels preview
Add an AirportDataSource class to parse Airport/Runway data from various
sources : currently X-Plane's apt.dat files, but can be extended.

Given a list of tiles, the AirportDataSource returns an AirportCollection
object (a hierarchy of Airport and Runway objects) representing all the
airports contained in given tiles and the immediate surroundings.

Given a zoomlevel, screen_res, fpa and fov, the AirportCollection can
then generate an appropriate zoom level zone.

The generated zones are optimized for texture usage : a ZLn texture will
not be included if it's almost all covered by ZLn+1 textures.

Currently, this feature is used when clicking the preview button in the
Custom_ZoomLevel window : the smart_zone_list function is called to
generate a custom zone_list from ZLmax to ZLmin.

This will later be integrated in the GUI as a separate feature instead.

There is also provision for using this feature in batch mode, but its
not yet integrated.